+++ ta4j-core/src/main/java/org/ta4j/core/criteria/ReturnOverMaxDrawdownCriterion.java
@@ -1,7 +1,7 @@
 /**
  * The MIT License (MIT)
  *
- * Copyright (c) 2017-2022 Ta4j Organization & respective
+ * Copyright (c) 2017-2023 Ta4j Organization & respective
  * authors (see AUTHORS)
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -33,7 +33,8 @@ import org.ta4j.core.num.Num;
 
 /**
  * Reward risk ratio criterion, defined as the {@link ReturnCriterion gross
- * return} over the {@link MaximumDrawdownCriterion maximum drawdown}.
+ * return} over the {@link MaximumDrawdownCriterion maximum drawdown}, returned
+ * in decimal format.
  */
 public class ReturnOverMaxDrawdownCriterion extends AbstractAnalysisCriterion {