+++ ta4j-core/src/main/java/org/ta4j/core/criteria/pnl/ProfitLossRatioCriterion.java
@@ -1,7 +1,7 @@
 /**
  * The MIT License (MIT)
  *
- * Copyright (c) 2017-2022 Ta4j Organization & respective
+ * Copyright (c) 2017-2023 Ta4j Organization & respective
  * authors (see AUTHORS)
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -31,7 +31,8 @@ import org.ta4j.core.num.Num;
 
 /**
  * Ratio gross profit and loss criterion = Average gross profit (includes
- * trading costs) / Average gross loss (includes trading costs).
+ * trading costs) / Average gross loss (includes trading costs), returned in
+ * decimal format.
  */
 public class ProfitLossRatioCriterion extends AbstractAnalysisCriterion {