+++ CHANGELOG.md
@@ -21,6 +21,9 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 ### Changed
 - **BarSeriesManager** consider finishIndex when running backtest
 - **BarSeriesManager** add `holdingTransaction
+- **BacktestExecutor** evaluates strategies in parallel when possible
+- **CachedIndicator** synchronize on getValue()
+
 
 ### Removed/Deprecated
 
@@ -33,6 +36,8 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - added `Indicator.getUnstableBars()`
 - added `TransformIndicator.pow()`
 - added `BarSeriesManager.getHoldingCostModel()` and `BarSeriesManager.getTransactionCostModel()`  to allow extending BarSeriesManager and reimplementing `run()`
+- added `MovingAverageCrossOverRangeBacktest.java` and `ETH-USD-PT5M-2023-3-13_2023-3-15.json` test data file to demonstrate parallel strategy evaluation
+- added javadoc improvements for percentage criteria
 
 ### Fixed
 - **Fixed** **CashFlow** fixed calculation with custom startIndex and endIndex